fixed user penalty data type

+++ insta_follow/settings.py
@@ -259,7 +259,7 @@ INSTAGRAM_CREDENTIALS = {
 DEVLYTIC_TOKEN = config('DEVLYTIC_TOKEN', default='')
 PUSH_API_URL = config('PUSH_API_URL', default='')
 
-USER_PENALTY_AMOUNT = config('USER_PENALTY_AMOUNT', default=1.5, cast=int)
+USER_PENALTY_AMOUNT = config('USER_PENALTY_AMOUNT', default=1.5, cast=float)
 
 FOLLOWER_LIMIT = config('FOLLOWER_LIMIT', default=20000, cast=int)
